# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Returns a Gallery API",
                "description": "API for accessing the Gallery",
                "operationId": "Get",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "gallery"
                ],
                "summary": "API Upload",
                "description": "API untuk mengirimkan data-data postingan",
                "operationId": "postGallery",
                "requestBody": {
                    "description": "Data untuk mengunggah gambar",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Judul Upload",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Deskripsi Gambar",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "File Gambar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}